+++ Cargo.toml
@@ -40,16 +40,13 @@ toml = "0.8.10"
 validator.workspace = true
 
 base = { path = "crates/base" }
-common = { path = "crates/common" }
 detect = { path = "crates/detect" }
 embedding = { path = "crates/embedding" }
-hnsw = { path = "crates/hnsw" }
 interprocess_atomic_wait = { path = "crates/interprocess-atomic-wait" }
 memfd = { path = "crates/memfd" }
 scopeguard = "1.2.0"
 send_fd = { path = "crates/send_fd" }
 service = { path = "crates/service" }
-storage = { path = "crates/storage" }
 
 [patch.crates-io]
 pgrx = { git = "https://github.com/tensorchord/pgrx.git", branch = "v0.12.0-alpha.1-patch" }
